backend/requirements.txt
flask
requests 
python-dotenv
pymongo